check if set is empty

+++ packages/core/database/lib/entity-manager/index.js
@@ -618,6 +618,10 @@ const createEntityManager = (db) => {
 
             const { idColumn, typeColumn } = morphColumn;
 
+            if (isEmpty(cleanRelationData.set)) {
+              continue;
+            }
+
             await this.createQueryBuilder(joinTable.name)
               .delete()
               .where({
